Fix gzip validations
